Added support for Oneof field for both Proto3 and proto2.
Also rework the generation of the other fields to ensure when
we generate multiple objects we have different values. Previous implementation
would always generate the same repeated object as the seed doesn't advance
instead was cached at the beginning of the execution. Added unit
test for these and new test verifying we indeed generating different
objects.